+++ packages/tree-extension/src/index.ts
@@ -170,6 +170,9 @@ const fileActions: JupyterFrontEndPlugin<void> = {
         selectionChanged.emit(void 0);
       };
     });
+    browser.model.pathChanged.connect(() => {
+      selectionChanged.emit(void 0);
+    });
 
     // Create a toolbar item that adds buttons to the file browser toolbar
     // to perform actions on the files
@@ -182,6 +185,58 @@ const fileActions: JupyterFrontEndPlugin<void> = {
     });
     for (const widget of fileActions.widgets) {
       toolbarRegistry.addFactory(FILE_BROWSER_FACTORY, widget.id, () => widget);
+    }
+  },
+};
+
+/**
+ * A plugin to set the default file browser settings.
+ */
+const fileBrowserSettings: JupyterFrontEndPlugin<void> = {
+  id: '@jupyter-notebook/tree-extension:settings',
+  description: 'Set up the default file browser settings',
+  requires: [IDefaultFileBrowser],
+  optional: [ISettingRegistry],
+  autoStart: true,
+  activate: (
+    app: JupyterFrontEnd,
+    browser: IDefaultFileBrowser,
+    settingRegistry: ISettingRegistry | null
+  ) => {
+    // Default config for notebook.
+    // This is a different set of defaults than JupyterLab.
+    const defaultFileBrowserConfig = {
+      navigateToCurrentDirectory: false,
+      singleClickNavigation: true,
+      showLastModifiedColumn: true,
+      showFileSizeColumn: true,
+      showHiddenFiles: false,
+      showFileCheckboxes: true,
+      sortNotebooksFirst: true,
+      showFullPath: false,
+    };
+
+    // Apply defaults on plugin activation
+    let key: keyof typeof defaultFileBrowserConfig;
+    for (key in defaultFileBrowserConfig) {
+      browser[key] = defaultFileBrowserConfig[key];
+    }
+
+    if (settingRegistry) {
+      void settingRegistry.load(FILE_BROWSER_PLUGIN_ID).then((settings) => {
+        function onSettingsChanged(settings: ISettingRegistry.ISettings): void {
+          let key: keyof typeof defaultFileBrowserConfig;
+          for (key in defaultFileBrowserConfig) {
+            const value = settings.get(key).user as boolean;
+            // only set the setting if it is defined by the user
+            if (value !== undefined) {
+              browser[key] = value;
+            }
+          }
+        }
+        settings.changed.connect(onSettingsChanged);
+        onSettingsChanged(settings);
+      });
     }
   },
 };
@@ -360,25 +415,6 @@ const notebookTreeWidget: JupyterFrontEndPlugin<INotebookTree> = {
       nbTreeWidget.tabBar.addTab(running.title);
     }
 
-    const settings = settingRegistry.load(FILE_BROWSER_PLUGIN_ID);
-    Promise.all([settings, app.restored])
-      .then(([settings]) => {
-        // Set Notebook 7 defaults if there is no user setting override
-        [
-          'showFileCheckboxes',
-          'showFileSizeColumn',
-          'sortNotebooksFirst',
-          'showFullPath',
-        ].forEach((setting) => {
-          if (settings.user[setting] === undefined) {
-            void settings.set(setting, true);
-          }
-        });
-      })
-      .catch((reason: Error) => {
-        console.error(reason.message);
-      });
-
     app.shell.add(nbTreeWidget, 'main', { rank: 100 });
 
     // add a separate tab for each setting editor
@@ -419,6 +455,7 @@ const notebookTreeWidget: JupyterFrontEndPlugin<INotebookTree> = {
 const plugins: JupyterFrontEndPlugin<any>[] = [
   createNew,
   fileActions,
+  fileBrowserSettings,
   fileFilterCommand,
   loadPlugins,
   openFileBrowser,